chore: refactor eth configuration
