+++ dozer-ingestion/src/connectors/ethereum/log/mod.rs
@@ -1,0 +1,7 @@
+mod helper;
+mod provider;
+mod sender;
+pub use provider::EthLogProvider;
+
+#[cfg(test)]
+mod tests;
