+++ dozer-ingestion/src/connectors/ethereum/log/tests/contracts/CustomEvent.json
