+++ dozer-ingestion/src/connectors/ethereum/mod.rs
@@ -1,7 +1,3 @@
-pub mod connector;
 pub mod helper;
-mod providers;
-mod sender;
-
-#[cfg(test)]
-mod tests;
+mod log;
+pub use log::EthLogProvider;
