--- dozer-ingestion/src/connectors/ethereum/providers/log.rs
@@ -1,1 +1,0 @@
-
